Refactored extension to make use of webpack
